requirements.txt
pygame==1.9.6
pyserial==3.4